requirements.txt
aiomysql==0.2.0
alembic==1.11.1
anyio==3.7.0
asyncmy==0.2.8
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==1.7.1
PyMySQL==1.0.3
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
